Update read_shape.R
Using st_set_crs just changes the crs text (as noted in the warning message of st_set_crs)  it doesn't transform it to the proposed 4326. This will actually move the spatial data. the shift is very very minimal in this case but the code is not correct as it is. 
ABS spatial data comes comes in either gda94 or gda2020 reference systems so need to be transformed into wgs84 (crs 4326) using st_transform.